Update CircularArraySARTSATraces.jl
Fix issue where state and actions would go out of sync when the capacity of the episode buffer is reached. The capacity of the AA' should be one less than that of the SS' multiplex trace as there is always one less action pushed compared to the state in experiments.